+++ packages/@lou.codes/utils/src/constructDataView.ts
@@ -1,0 +1,8 @@
+import { construct } from "./construct.js";
+
+/**
+ * Functional alias for `new DataView()`.
+ *
+ * @category Classes
+ */
+export const constructDataView = construct(DataView);