+++ packages/@lou.codes/utils/src/constructTextEncoder.ts
@@ -1,0 +1,8 @@
+import { construct } from "./construct.js";
+
+/**
+ * Functional alias for `new TextEncoder()`.
+ *
+ * @category Classes
+ */
+export const constructTextEncoder = construct(TextEncoder);