+++ packages/@lou.codes/utils/src/toString.ts
@@ -1,21 +1,16 @@
 import type { Strigifiable } from "@lou.codes/types";
 
 /**
- * Parses given value to string, if number is given, then that's used as a radix
- * and a function is returned expecting an actual value.
+ * Parses given value to string.
  *
  * @category Common
  * @example
  * ```typescript
- * const toDecimal = toString(10);
- *
- * toDecimal(10); // "10"
- * toString(16)(16); // "10"
- * toString(2)(2); // "10"
+ * toString(10); // "10"
  * toString({}); // "[object Object]"
  * toString({ toString: _ => "test" }); // "test"
  * ```
- * @returns string or function expecting a value.
+ * @returns stringified value.
  */
 export const toString = <
 	const Input extends Strigifiable | { readonly toString: () => string },