+++ packages/@lou.codes/utils/tests/constructTextEncoder.test.ts
@@ -1,0 +1,9 @@
+import type { Test } from "@lou.codes/test";
+import { constructTextEncoder } from "../src/constructTextEncoder.js";
+
+export const constructTextEncoderTest = {
+	given: "a constructTextEncoder",
+	must: "return new instance of TextEncoder",
+	received: () => constructTextEncoder().constructor.name,
+	wanted: () => "TextEncoder",
+} satisfies Test<string>;